requirements.txt
setuptools==75.8.0
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
tomli==2.2.1
PyYAML==6.0.2
numba==0.60.0
line_profiler==4.2.0
numpy==2.0.2
scipy==1.13.1
llvmlite==0.43.0
